feat: block syncer add dual db support config (#32)
* feat: dual db feature support

* fix:statement policy id & accountId data type

---------
